requirements.txt
Django~=3.2.9
pip~=21.3.1
SQLAlchemy~=1.4.27
cx-Oracle~=8.3.0
alembic~=1.7.5
Mako~=1.1.6
djangorestframework~=3.12.4